requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
click==6.7
cryptography==2.3.1
dnspython==1.15.0
eventlet==0.24.1
Flask==1.0.2
flask-marshmallow==0.9.0
Flask-MQTT==1.0.4
Flask-SocketIO==3.0.1
Flask-SQLAlchemy==2.3.2
greenlet==0.4.14
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
marshmallow==2.15.4
marshmallow-sqlalchemy==0.14.1
monotonic==1.5
paho-mqtt==1.4.0
pycparser==2.18
PyMySQL==0.9.2
python-dotenv==0.9.1
python-engineio==2.2.0
python-socketio==2.0.0
six==1.11.0
SQLAlchemy==1.2.11
typing==3.6.6
virtualenv==16.0.0
virtualenvwrapper-win==1.2.5
Werkzeug==0.14.1
